fix inconsistency in CV toString() method
Originally, the `personalStatement` field in the `CV` class would not
appear in the `toString()` output if it had no been set. This was
inconsistent with the `toString()` method in other CV document
classes.

Have changed the `toString()` method to always output the
`personalStatement` field. It will now output a value of `null` if
the field has not been set.

cvg-6